Updated way headwater and local inflow temperature is calculated
Update to establish consistency in the way the temperature for headwater inflow and distributed inflow is calculated. In both cases, the final water temperature for local inflow is now based on the mixing ratios for local baseflow, runoff from rainfall, and runoff from snow and glacier melt.

For baseflow, water temperature is assumed equal to the soil temperature, with four options:
  * Mohseni et al. regression on air temperature as proxy for soil temperature
  * Pseudo-linear relation on air temperature as proxy for soil temperature
  * Smoothed air temperature as estimate of soil temperature
  * Soil temperature as direct forcing/input

The temperature of surface runoff from rainfall is assumed equal to the air temperature, and the temperature of runoff from snow and glacier melt is assumed equal to T_snow (which is typically set to 0oC).